requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2
qiskit>=1.0.0
qiskit-aer>=0.14.0
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
